Simplify: Remove useForm and move logics to Form component
